Fix ToC Generation
Explanation:

The documentation is using middleman-blog

Each page is encapsulated in a Middleman::Blog::BlogArticle structure
which override the `render` method.

If the layout is `nil`, it looks for a default layout. So obviously,
even when `body_for` was looking for a nil layout, a layout was applied,
and it created an infinite loop, as this layout required also a ToC

I've just override the problematic method