Unsupported erlang OTP-24.*
Adding a note on unsupported OTP-24.*  (on scalingo-18) at the moment. The working versions on this stack are listing there -> https://github.com/HashNuke/heroku-buildpack-elixir-otp-builds/blob/master/otp-versions. They are aware of the issue (related to builds), no fix yet (if ever since the *-20 stacks from Heroku and Gigalixir are supporting those versions).